Upgrade to TypeScript
