+++ decs.d.ts
@@ -1,0 +1,2 @@
+declare module "Xvfb"
+declare module "chromium"
