+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "puppeteer-real-browser",
-  "version": "1.1.5",
+  "version": "1.1.6",
   "description": "This package is designed to bypass puppeteer's bot-detecting captchas such as Cloudflare. It acts like a real browser and can be managed with puppeteer.",
   "type": "module",
   "main": "index.js",
@@ -32,5 +32,9 @@
     "url": "https://github.com/zfcsoftware/puppeteer-real-browser"
   },
   "readme": "README.md",
-  "homepage": "https://github.com/zfcsoftware/puppeteer-real-browser"
+  "homepage": "https://github.com/zfcsoftware/puppeteer-real-browser",
+  "devDependencies": {
+    "@types/chrome-remote-interface": "^0.31.14",
+    "@types/cli-color": "^2.0.6"
+  }
 }
