+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+  	"compilerOptions": {
+    	"target": "es2016",
+    	"module": "commonjs",
+    	"esModuleInterop": true,
+    	"forceConsistentCasingInFileNames": true,
+    	"strict": true,
+    	"skipLibCheck": true
+  	},
+	"include": [
+		"decs.d.ts"
+	]
+}
